Update to 'spell' class + revisisons
- Magic class
+ Spell class as base
+ Enchantment as base
= Necessary revisions to include new classes
