+++ Enchantment.cpp
@@ -1,0 +1,5 @@
+#include "Enchantment.hpp"
+
+Enchantment::Enchantment() {
+
+}
