+++ Enchantment.hpp
@@ -1,0 +1,14 @@
+#ifndef ENCHANTMENT_HPP
+#define ENCHANTMENT_HPP
+
+class Item;
+
+//! Enchantment class
+/*! Derived from Magic; Applies bonus stats to weapons and armor. */
+class Enchantment {
+
+public:
+  Enchantment(); //!< Default Constructor
+};
+
+#endif
