--- Magic.cpp
@@ -1,17 +1,0 @@
-#include "Magic.hpp"
-
-Magic::Magic(){
-
-}
-
-Spell::Spell(){
-
-}
-
-Buff::Buff(){
-
-}
-
-Enchantment::Enchantment(){
-
-}
