+++ Spell.cpp
@@ -1,0 +1,9 @@
+#include "Spell.hpp"
+
+Spell::Spell(){
+
+}
+
+Buff::Buff(){
+
+}
